Fix Cluster Settings > Cluster Operators sort by status
